eliminating warnings, mostly with regards to signed vs unsigned comparisons, some of which could have been functional bugs if negative values were used (for offsets, etc). cleaned up a couple of macros and cleaned up some ifdefs.

+++ libclamav/tnef.c
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2015 Cisco Systems, Inc. and/or its affiliates. All rights reserved.
+ *  Copyright (C) 2015, 2017 Cisco Systems, Inc. and/or its affiliates. All rights reserved.
  *  Copyright (C) 2007-2008 Sourcefire, Inc.
  *
  *  Authors: Nigel Horne
@@ -265,7 +265,7 @@ tnef_message(fmap_t *map, off_t *pos, uint16_t type, uint16_t tag, int32_t lengt
 
 	/*cli_dbgmsg("%lu %lu\n", (long)(offset + length), ftell(fp));*/
 
-	if(!CLI_ISCONTAINED2(0, fsize, (off_t)offset, (off_t)length)) {
+	if(!CLI_ISCONTAINED2(0, fsize, offset, (off_t)length)) {
 		cli_dbgmsg("TNEF: Incorrect length field in tnef_message\n");
 		return -1;
 	}
